backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Limiter
psycopg2-binary
python-dotenv
Flask-Migrate
phonenumbers
email-validator
python-magic
flask-cors
python-chess
Flask-SocketIO
eventlet
requests